+++ Utilidades/Utilidades.h
@@ -5,7 +5,7 @@ int comprobacionTLF(int TLF);
 int comprobacionIBAN(char *IBAN);
 int ponerMayusculas(char * cadena);
 int comprobacionEmail(char *email);
-int comprobacionN_SS(int n_ss);
+int comprobacionN_SS(unsigned long long n_ss);
 int comprobacionSueldo(double sueldo);
 int comprobacionHorario (char *horario);
 #endif
